refactor: rewritten types inside .ts files
