--- types/index.d.ts
@@ -1,4 +1,0 @@
-import './global'
-
-export * from './strapi'
-export * from './module'
